requirements_old.txt
ansi2html==1.9.2
antlr4-python3-runtime==4.9.3
attrs==23.2.0
basemap==1.4.1
basemap-data==1.3.2
bokeh==3.5.1
cachelib==0.9.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorcet==3.1.0
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.7.1
dask-expr==1.1.9
datashader==0.16.3
distributed==2024.7.1
fastparquet==2024.5.0
filelock==3.15.4
fiona==1.9.6
Flask==2.2.5
Flask-Caching==2.3.0
fonttools==4.53.1
fsspec==2024.6.1
geopandas==1.0.1
gunicorn==22.0.0
huggingface-hub==0.24.6
hydra-core==1.3.2
idna==3.7
imbalanced-learn==0.8.1
importlib_metadata==8.2.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
llvmlite==0.43.0
locket==1.0.0
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
msgpack==1.0.8
multipledispatch==1.0.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
packaging==23.2
pandas==2.0.3
param==2.1.1
partd==1.4.2
patsy==0.5.6
pillow==10.4.0
plotly==5.23.0
portalocker==2.10.1
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyct==0.5.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
retrying==1.3.4
# Editable install with no version control (SAM-2==1.0)
-e /Users/cobi/PycharmProjects/unlimited-analytics/segment-anything-2
scikit-learn==1.5.1
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sortedcontainers==2.4.0
statsmodels==0.14.2
sympy==1.13.2
tblib==3.0.0
tenacity==9.0.0
threadpoolctl==3.5.0
toolz==0.12.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==2.2.3
xarray==2024.7.0
xyzservices==2024.6.0
zict==3.0.0
zipp==3.19.2
